requirements.txt
# Core dependencies
streamlit>=1.50.0
python-dotenv>=1.1.1

# Document processing
pdfplumber>=0.10.0
docx2txt>=0.8

# Language detection
langdetect>=1.0.9

# Multilingual embeddings and NLP
sentence-transformers>=2.2.2
transformers>=4.35.0
torch>=2.0.0

# Vector database
qdrant-client>=1.7.0

# HTTP requests
requests>=2.31.0

# Additional utilities
numpy>=1.24.0
pandas>=2.0.0
nltk>=3.8.1

# Optional: For better text processing
indic-nlp-library>=0.81

# OCR (for image-based PDFs)
pytesseract>=0.3.10
pdf2image>=1.17.0
Pillow>=10.0.0
